+++ chrome_target.go
@@ -1,7 +1,7 @@
 /*
 The MIT License (MIT)
 
-Copyright (c) 2017 isaac dawson
+Copyright (c) 2019 isaac dawson
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -58,8 +58,8 @@ type TargetInfo struct {
 // Events are handled by mapping the method name to a function which takes a target and byte output.
 // For now, callers will need to unmarshall the types themselves.
 type ChromeTarget struct {
-	sendId      int64                    		       // An Id which is atomically incremented per request.
-							       // must be at top because of alignement and atomic usage 
+	sendId int64 // An Id which is atomically incremented per request.
+	// must be at top because of alignement and atomic usage
 	replyLock       sync.RWMutex                           // lock for dispatching responses
 	replyDispatcher map[int64]chan *gcdmessage.Message     // Replies to synch methods using a non-buffered channel
 	eventLock       sync.RWMutex                           // lock for dispatching events
@@ -71,7 +71,9 @@ type ChromeTarget struct {
 	Animation            *gcdapi.Animation
 	ApplicationCache     *gcdapi.ApplicationCache // application cache API
 	Browser              *gcdapi.Browser
+	BackgroundService    *gcdapi.BackgroundService
 	CacheStorage         *gcdapi.CacheStorage
+	Cast                 *gcdapi.Cast
 	Console              *gcdapi.Console           // console API
 	CSS                  *gcdapi.CSS               // CSS API
 	Database             *gcdapi.Database          // Database API
@@ -142,7 +144,9 @@ func (c *ChromeTarget) Init() {
 	c.Animation = gcdapi.NewAnimation(c)
 	c.ApplicationCache = gcdapi.NewApplicationCache(c)
 	c.Browser = gcdapi.NewBrowser(c)
+	c.BackgroundService = gcdapi.NewBackgroundService(c)
 	c.CacheStorage = gcdapi.NewCacheStorage(c)
+	c.Cast = gcdapi.NewCast(c)
 	c.Console = gcdapi.NewConsole(c)
 	c.CSS = gcdapi.NewCSS(c)
 	c.Database = gcdapi.NewDatabase(c)