+++ gcdapi/emulation.go
@@ -15,22 +15,6 @@ type EmulationScreenOrientation struct {
 	Angle int    `json:"angle"` // Orientation angle.
 }
 
-// Notification sent after the virtual time has advanced.
-type EmulationVirtualTimeAdvancedEvent struct {
-	Method string `json:"method"`
-	Params struct {
-		VirtualTimeElapsed float64 `json:"virtualTimeElapsed"` // The amount of virtual time that has elapsed in milliseconds since virtual time was first enabled.
-	} `json:"Params,omitempty"`
-}
-
-// Notification sent after the virtual time has paused.
-type EmulationVirtualTimePausedEvent struct {
-	Method string `json:"method"`
-	Params struct {
-		VirtualTimeElapsed float64 `json:"virtualTimeElapsed"` // The amount of virtual time that has elapsed in milliseconds since virtual time was first enabled.
-	} `json:"Params,omitempty"`
-}
-
 type Emulation struct {
 	target gcdmessage.ChromeTargeter
 }