+++ gcdapi/input.go
@@ -129,12 +129,16 @@ type InputDispatchMouseEventParams struct {
 	Timestamp float64 `json:"timestamp,omitempty"`
 	// Mouse button (default: "none").
 	Button string `json:"button,omitempty"`
+	// A number indicating which buttons are pressed on the mouse when a mouse event is triggered. Left=1, Right=2, Middle=4, Back=8, Forward=16, None=0.
+	Buttons int `json:"buttons,omitempty"`
 	// Number of times the mouse button was clicked (default: 0).
 	ClickCount int `json:"clickCount,omitempty"`
 	// X delta in CSS pixels for mouse wheel event (default: 0).
 	DeltaX float64 `json:"deltaX,omitempty"`
 	// Y delta in CSS pixels for mouse wheel event (default: 0).
 	DeltaY float64 `json:"deltaY,omitempty"`
+	// Pointer type (default: "mouse").
+	PointerType string `json:"pointerType,omitempty"`
 }
 
 // DispatchMouseEventWithParams - Dispatches a mouse event to the page.
@@ -149,10 +153,12 @@ func (c *Input) DispatchMouseEventWithParams(v *InputDispatchMouseEventParams) (
 // modifiers - Bit field representing pressed modifier keys. Alt=1, Ctrl=2, Meta/Command=4, Shift=8 (default: 0).
 // timestamp - Time at which the event occurred.
 // button - Mouse button (default: "none").
+// buttons - A number indicating which buttons are pressed on the mouse when a mouse event is triggered. Left=1, Right=2, Middle=4, Back=8, Forward=16, None=0.
 // clickCount - Number of times the mouse button was clicked (default: 0).
 // deltaX - X delta in CSS pixels for mouse wheel event (default: 0).
 // deltaY - Y delta in CSS pixels for mouse wheel event (default: 0).
-func (c *Input) DispatchMouseEvent(theType string, x float64, y float64, modifiers int, timestamp float64, button string, clickCount int, deltaX float64, deltaY float64) (*gcdmessage.ChromeResponse, error) {
+// pointerType - Pointer type (default: "mouse").
+func (c *Input) DispatchMouseEvent(theType string, x float64, y float64, modifiers int, timestamp float64, button string, buttons int, clickCount int, deltaX float64, deltaY float64, pointerType string) (*gcdmessage.ChromeResponse, error) {
 	var v InputDispatchMouseEventParams
 	v.TheType = theType
 	v.X = x
@@ -160,9 +166,11 @@ func (c *Input) DispatchMouseEvent(theType string, x float64, y float64, modifie
 	v.Modifiers = modifiers
 	v.Timestamp = timestamp
 	v.Button = button
+	v.Buttons = buttons
 	v.ClickCount = clickCount
 	v.DeltaX = deltaX
 	v.DeltaY = deltaY
+	v.PointerType = pointerType
 	return c.DispatchMouseEventWithParams(&v)
 }