+++ gcdapi/network.go
@@ -114,7 +114,7 @@ type NetworkWebSocketResponse struct {
 type NetworkWebSocketFrame struct {
 	Opcode      float64 `json:"opcode"`      // WebSocket message opcode.
 	Mask        bool    `json:"mask"`        // WebSocket message mask.
-	PayloadData string  `json:"payloadData"` // WebSocket message payload data.
+	PayloadData string  `json:"payloadData"` // WebSocket message payload data. If the opcode is 1, this is a text message and payloadData is a UTF-8 string. If the opcode isn't 1, then payloadData is a base64 encoded string representing binary data.
 }
 
 // Information about the cached resource.
@@ -198,7 +198,6 @@ type NetworkSignedExchangeSignature struct {
 // Information about a signed exchange header. https://wicg.github.io/webpackage/draft-yasskin-httpbis-origin-signed-exchanges-impl.html#cbor-representation
 type NetworkSignedExchangeHeader struct {
 	RequestUrl      string                            `json:"requestUrl"`      // Signed exchange request URL.
-	RequestMethod   string                            `json:"requestMethod"`   // Signed exchange request method.
 	ResponseCode    int                               `json:"responseCode"`    // Signed exchange response code.
 	ResponseHeaders map[string]interface{}            `json:"responseHeaders"` // Signed exchange response headers.
 	Signatures      []*NetworkSignedExchangeSignature `json:"signatures"`      // Signed exchange response signature.